Round values for better visual

+++ temp_sense/api/management/commands/discord_devices_without_readings.py
@@ -17,7 +17,7 @@ def get_remaining_tokens_and_days():
 
     total_available_tokens = settings.TOTAL_AVAILABLE_TOKENS - total_tokens_used
 
-    return round(total_available_tokens, 0), round(days_left_of_tokens, 0)
+    return int(round(total_available_tokens, 0)), round(days_left_of_tokens, 0)
 
 
 class Command(BaseCommand):
